+++ enrolment/tests/test_helpers.py
@@ -103,24 +103,6 @@ def test_user_has_verified_company_404():
     assert helpers.user_has_verified_company(sso_user_id=1) is False
 
 
-def test_get_employees_label():
-    assert helpers.get_employees_label('1001-10000') == '1,001-10,000'
-
-
-def test_get_sectors_labels():
-    values = ['AGRICULTURE_HORTICULTURE_AND_FISHERIES', 'AEROSPACE']
-    expected = ['Agriculture, horticulture and fisheries', 'Aerospace']
-    assert helpers.get_sectors_labels(values) == expected
-
-
-def test_get_employees_label_none():
-    assert helpers.get_employees_label('') == ''
-
-
-def test_get_sectors_labels_none():
-    assert helpers.get_sectors_labels([]) == []
-
-
 @patch.object(helpers, 'get_companies_house_profile')
 def test_store_companies_house_profile_in_session_saves_in_session(
     mock_get_companies_house_profile, client
